+++ APIRestaurante.Domain/Entities/MenuItem.cs
@@ -4,6 +4,6 @@ namespace APIRestaurante.Domain.Entities
     {
         public int Id { get; set; }
         public required string Name { get; set; }
-        public required decimal PriceCents { get; set; }
+        public required int PriceCents { get; set; }
     }
 }
